Revert "Feat/condo/doma 5651/read news by user (#3138)"
This reverts commit 70b084c91f2d2647145ceba8a606a1f6a53d46dc.
