+++ apps/condo/domains/news/access/NewsItem.js
@@ -2,12 +2,11 @@
  * Generated by `createschema news.OrganizationNewsItem 'organization:Relationship:Organization:CASCADE; title:Text; body:Text; type:Select:common,emergency'`
  */
 
-const { compact, get, isEmpty, uniq } = require('lodash')
+const { compact, flatten,  get, isEmpty, uniq } = require('lodash')
 
 const { throwAuthenticationError } = require('@open-condo/keystone/apolloErrorFormatter')
 const { find, getById } = require('@open-condo/keystone/schema')
 
-const { queryFindNewsItemsScopesByResidents } = require('@condo/domains/news/utils/accessSchema')
 const {
     queryOrganizationEmployeeFor,
     queryOrganizationEmployeeFromRelatedOrganizationFor,
@@ -26,11 +25,62 @@ async function canReadNewsItems ({ authentication: { item: user } }) {
 
         if (isEmpty(residents)) return false
 
-        const scopesConditions = queryFindNewsItemsScopesByResidents(residents)
+        const organizationsIds = uniq(compact(residents.map(resident => get(resident, 'organization'))))
+
+        const scopesConditions = {
+            deletedAt: null,
+            OR: flatten(residents.map((resident) => {
+                const organizationId = get(resident, 'organization')
+                const propertyId = get(resident, 'property')
+                const unitType = get(resident, 'unitType')
+                const unitName = get(resident, 'unitName')
+
+                return [
+                    {
+                        AND: [
+                            {
+                                newsItem: { organization: { id: organizationId } },
+                            },
+                            {
+                                OR: [
+                                    {
+                                        AND: [
+                                            { property_is_null: true },
+                                            { unitType: null },
+                                            { unitName: null },
+                                        ],
+                                    },
+                                    {
+                                        AND: [
+                                            { property: { id: propertyId } },
+                                            { unitType: null },
+                                            { unitName: null },
+                                        ],
+                                    },
+                                    {
+                                        AND: [
+                                            { property: { id: propertyId } },
+                                            { unitType },
+                                            { unitName: null },
+                                        ],
+                                    },
+                                    {
+                                        AND: [
+                                            { property: { id: propertyId } },
+                                            { unitType },
+                                            { unitName: unitName },
+                                        ],
+                                    },
+                                ],
+                            },
+                        ],
+                    },
+                ]
+            })),
+        }
+
         const newsItemsScopes = await find('NewsItemScope', scopesConditions)
         const scopedNewsItemsIds = uniq(compact(newsItemsScopes.map((newsItemScope) => get(newsItemScope, 'newsItem'))))
-
-        const organizationsIds = uniq(compact(residents.map((resident) => get(resident, 'organization'))))
 
         return {
             id_in: scopedNewsItemsIds,
