+++ apps/condo/domains/news/constants/errors.js
@@ -1,7 +1,5 @@
 const EMPTY_VALID_BEFORE_DATE = 'EMPTY_VALID_BEFORE_DATE'
-const NEWS_ITEM_ALREADY_READ = 'NEWS_ITEM_ALREADY_READ'
 
 module.exports = {
     EMPTY_VALID_BEFORE_DATE,
-    NEWS_ITEM_ALREADY_READ,
 }
