--- apps/condo/domains/news/utils/clientSchema/NewsItemUserRead.ts
@@ -1,32 +1,0 @@
-/**
- * Generated by `createschema news.NewsItemUserRead 'newsItem:Relationship:NewsItem:CASCADE; user:Relationship:User:CASCADE'`
- */
-
-import {
-    NewsItemUserRead,
-    NewsItemUserReadCreateInput,
-    NewsItemUserReadUpdateInput,
-    QueryAllNewsItemUserReadsArgs,
-} from '@app/condo/schema'
-
-import { generateReactHooks } from '@open-condo/codegen/generate.hooks'
-
-import { NewsItemUserRead as NewsItemUserReadGQL } from '@condo/domains/news/gql'
-
-// TODO(codegen): write utils like convertToFormState and formValuesProcessor if needed, otherwise delete this TODO
-
-const {
-    useObject,
-    useObjects,
-    useCreate,
-    useUpdate,
-    useSoftDelete,
-} = generateReactHooks<NewsItemUserRead, NewsItemUserReadCreateInput, NewsItemUserReadUpdateInput, QueryAllNewsItemUserReadsArgs>(NewsItemUserReadGQL)
-
-export {
-    useObject,
-    useObjects,
-    useCreate,
-    useUpdate,
-    useSoftDelete,
-}
