+++ apps/condo/domains/news/utils/testSchema/index.js
@@ -11,13 +11,11 @@ const { NewsItem: NewsItemGQL } = require('@condo/domains/news/gql')
 const { NEWS_TYPE_COMMON } = require('@condo/domains/news/constants/newsTypes')
 const { NewsItemScope: NewsItemScopeGQL } = require('@condo/domains/news/gql')
 const { NewsItemTemplate: NewsItemTemplateGQL } = require('@condo/domains/news/gql')
-const { NewsItemUserRead: NewsItemUserReadGQL } = require('@condo/domains/news/gql')
 /* AUTOGENERATE MARKER <IMPORT> */
 
 const NewsItem = generateGQLTestUtils(NewsItemGQL)
 const NewsItemScope = generateGQLTestUtils(NewsItemScopeGQL)
 const NewsItemTemplate = generateGQLTestUtils(NewsItemTemplateGQL)
-const NewsItemUserRead = generateGQLTestUtils(NewsItemUserReadGQL)
 /* AUTOGENERATE MARKER <CONST> */
 
 async function createTestNewsItem (client, organization, extraAttrs = {}) {
@@ -122,43 +120,11 @@ async function updateTestNewsItemTemplate (client, id, extraAttrs = {}) {
     return [obj, attrs]
 }
 
-async function createTestNewsItemUserRead (client, newsItem, user, extraAttrs = {}) {
-    if (!client) throw new Error('no client')
-    if (!newsItem || !newsItem.id) throw new Error('no newsItem.id')
-    if (!user || !user.id) throw new Error('no user.id')
-    const sender = { dv: 1, fingerprint: faker.random.alphaNumeric(8) }
-
-    const attrs = {
-        dv: 1,
-        sender,
-        newsItem: { connect: { id: newsItem.id } },
-        user: { connect: { id: user.id } },
-        ...extraAttrs,
-    }
-    const obj = await NewsItemUserRead.create(client, attrs)
-    return [obj, attrs]
-}
-
-async function updateTestNewsItemUserRead (client, id, extraAttrs = {}) {
-    if (!client) throw new Error('no client')
-    if (!id) throw new Error('no id')
-    const sender = { dv: 1, fingerprint: faker.random.alphaNumeric(8) }
-
-    const attrs = {
-        dv: 1,
-        sender,
-        ...extraAttrs,
-    }
-    const obj = await NewsItemUserRead.update(client, id, attrs)
-    return [obj, attrs]
-}
-
 /* AUTOGENERATE MARKER <FACTORY> */
 
 module.exports = {
     NewsItem, createTestNewsItem, updateTestNewsItem,
     NewsItemScope, createTestNewsItemScope, updateTestNewsItemScope,
     NewsItemTemplate, createTestNewsItemTemplate, updateTestNewsItemTemplate,
-    NewsItemUserRead, createTestNewsItemUserRead, updateTestNewsItemUserRead,
 /* AUTOGENERATE MARKER <EXPORTS> */
 }
