+++ apps/condo/schema.ts
@@ -29237,30 +29237,6 @@ export type Mutation = {
   deleteNewsItemTemplate?: Maybe<NewsItemTemplate>;
   /**  Delete multiple NewsItemTemplate items by ID.  */
   deleteNewsItemTemplates?: Maybe<Array<Maybe<NewsItemTemplate>>>;
-  /**  Create a single NewsItemUserReadHistoryRecord item.  */
-  createNewsItemUserReadHistoryRecord?: Maybe<NewsItemUserReadHistoryRecord>;
-  /**  Create multiple NewsItemUserReadHistoryRecord items.  */
-  createNewsItemUserReadHistoryRecords?: Maybe<Array<Maybe<NewsItemUserReadHistoryRecord>>>;
-  /**  Update a single NewsItemUserReadHistoryRecord item by ID.  */
-  updateNewsItemUserReadHistoryRecord?: Maybe<NewsItemUserReadHistoryRecord>;
-  /**  Update multiple NewsItemUserReadHistoryRecord items by ID.  */
-  updateNewsItemUserReadHistoryRecords?: Maybe<Array<Maybe<NewsItemUserReadHistoryRecord>>>;
-  /**  Delete a single NewsItemUserReadHistoryRecord item by ID.  */
-  deleteNewsItemUserReadHistoryRecord?: Maybe<NewsItemUserReadHistoryRecord>;
-  /**  Delete multiple NewsItemUserReadHistoryRecord items by ID.  */
-  deleteNewsItemUserReadHistoryRecords?: Maybe<Array<Maybe<NewsItemUserReadHistoryRecord>>>;
-  /**  Create a single NewsItemUserRead item.  */
-  createNewsItemUserRead?: Maybe<NewsItemUserRead>;
-  /**  Create multiple NewsItemUserRead items.  */
-  createNewsItemUserReads?: Maybe<Array<Maybe<NewsItemUserRead>>>;
-  /**  Update a single NewsItemUserRead item by ID.  */
-  updateNewsItemUserRead?: Maybe<NewsItemUserRead>;
-  /**  Update multiple NewsItemUserRead items by ID.  */
-  updateNewsItemUserReads?: Maybe<Array<Maybe<NewsItemUserRead>>>;
-  /**  Delete a single NewsItemUserRead item by ID.  */
-  deleteNewsItemUserRead?: Maybe<NewsItemUserRead>;
-  /**  Delete multiple NewsItemUserRead items by ID.  */
-  deleteNewsItemUserReads?: Maybe<Array<Maybe<NewsItemUserRead>>>;
   /**  Create a single WebhookHistoryRecord item.  */
   createWebhookHistoryRecord?: Maybe<WebhookHistoryRecord>;
   /**  Create multiple WebhookHistoryRecord items.  */
@@ -37059,68 +37035,6 @@ export type MutationDeleteNewsItemTemplatesArgs = {
 };
 
 
-export type MutationCreateNewsItemUserReadHistoryRecordArgs = {
-  data?: Maybe<NewsItemUserReadHistoryRecordCreateInput>;
-};
-
-
-export type MutationCreateNewsItemUserReadHistoryRecordsArgs = {
-  data?: Maybe<Array<Maybe<NewsItemUserReadHistoryRecordsCreateInput>>>;
-};
-
-
-export type MutationUpdateNewsItemUserReadHistoryRecordArgs = {
-  id: Scalars['ID'];
-  data?: Maybe<NewsItemUserReadHistoryRecordUpdateInput>;
-};
-
-
-export type MutationUpdateNewsItemUserReadHistoryRecordsArgs = {
-  data?: Maybe<Array<Maybe<NewsItemUserReadHistoryRecordsUpdateInput>>>;
-};
-
-
-export type MutationDeleteNewsItemUserReadHistoryRecordArgs = {
-  id: Scalars['ID'];
-};
-
-
-export type MutationDeleteNewsItemUserReadHistoryRecordsArgs = {
-  ids?: Maybe<Array<Scalars['ID']>>;
-};
-
-
-export type MutationCreateNewsItemUserReadArgs = {
-  data?: Maybe<NewsItemUserReadCreateInput>;
-};
-
-
-export type MutationCreateNewsItemUserReadsArgs = {
-  data?: Maybe<Array<Maybe<NewsItemUserReadsCreateInput>>>;
-};
-
-
-export type MutationUpdateNewsItemUserReadArgs = {
-  id: Scalars['ID'];
-  data?: Maybe<NewsItemUserReadUpdateInput>;
-};
-
-
-export type MutationUpdateNewsItemUserReadsArgs = {
-  data?: Maybe<Array<Maybe<NewsItemUserReadsUpdateInput>>>;
-};
-
-
-export type MutationDeleteNewsItemUserReadArgs = {
-  id: Scalars['ID'];
-};
-
-
-export type MutationDeleteNewsItemUserReadsArgs = {
-  ids?: Maybe<Array<Scalars['ID']>>;
-};
-
-
 export type MutationCreateWebhookHistoryRecordArgs = {
   data?: Maybe<WebhookHistoryRecordCreateInput>;
 };
@@ -38517,313 +38431,6 @@ export type NewsItemUpdateInput = {
   newId?: Maybe<Scalars['String']>;
   dv?: Maybe<Scalars['Int']>;
   sender?: Maybe<SenderFieldInput>;
-};
-
-/**  The fact the user has read the particular news item  */
-export type NewsItemUserRead = {
-  __typename?: 'NewsItemUserRead';
-  /**
-   * This virtual field will be resolved in one of the following ways (in this order):
-   *  1. Execution of 'labelResolver' set on the NewsItemUserRead List config, or
-   *  2. As an alias to the field set on 'labelField' in the NewsItemUserRead List config, or
-   *  3. As an alias to a 'name' field on the NewsItemUserRead List (if one exists), or
-   *  4. As an alias to the 'id' field on the NewsItemUserRead List.
-   */
-  _label_?: Maybe<Scalars['String']>;
-  /**  The news item the user has read  */
-  newsItem?: Maybe<NewsItem>;
-  /**  The user who has read the particular news item  */
-  user?: Maybe<User>;
-  id: Scalars['ID'];
-  v?: Maybe<Scalars['Int']>;
-  createdAt?: Maybe<Scalars['String']>;
-  updatedAt?: Maybe<Scalars['String']>;
-  /**  Identifies a user, which has created this record. It is a technical connection, that can represent real users, as well as automated systems (bots, scripts). This field should not participate in business logic.  */
-  createdBy?: Maybe<User>;
-  /**  Identifies a user, which has updated this record. It is a technical connection, that can represent real users, as well as automated systems (bots, scripts). This field should not participate in business logic.  */
-  updatedBy?: Maybe<User>;
-  deletedAt?: Maybe<Scalars['String']>;
-  newId?: Maybe<Scalars['String']>;
-  /**  Data structure Version  */
-  dv?: Maybe<Scalars['Int']>;
-  /**  Client-side device identification used for the anti-fraud detection. Example `{ dv: 1, fingerprint: 'VaxSw2aXZa'}`. Where the `fingerprint` should be the same for the same devices and it's not linked to the user ID. It's the device ID like browser / mobile application / remote system  */
-  sender?: Maybe<SenderField>;
-};
-
-export type NewsItemUserReadCreateInput = {
-  newsItem?: Maybe<NewsItemRelateToOneInput>;
-  user?: Maybe<UserRelateToOneInput>;
-  v?: Maybe<Scalars['Int']>;
-  createdAt?: Maybe<Scalars['String']>;
-  updatedAt?: Maybe<Scalars['String']>;
-  createdBy?: Maybe<UserRelateToOneInput>;
-  updatedBy?: Maybe<UserRelateToOneInput>;
-  deletedAt?: Maybe<Scalars['String']>;
-  newId?: Maybe<Scalars['String']>;
-  dv?: Maybe<Scalars['Int']>;
-  sender?: Maybe<SenderFieldInput>;
-};
-
-/**  A keystone list  */
-export type NewsItemUserReadHistoryRecord = {
-  __typename?: 'NewsItemUserReadHistoryRecord';
-  /**
-   * This virtual field will be resolved in one of the following ways (in this order):
-   *  1. Execution of 'labelResolver' set on the NewsItemUserReadHistoryRecord List config, or
-   *  2. As an alias to the field set on 'labelField' in the NewsItemUserReadHistoryRecord List config, or
-   *  3. As an alias to a 'name' field on the NewsItemUserReadHistoryRecord List (if one exists), or
-   *  4. As an alias to the 'id' field on the NewsItemUserReadHistoryRecord List.
-   */
-  _label_?: Maybe<Scalars['String']>;
-  newsItem?: Maybe<Scalars['String']>;
-  user?: Maybe<Scalars['String']>;
-  id: Scalars['ID'];
-  v?: Maybe<Scalars['Int']>;
-  createdAt?: Maybe<Scalars['String']>;
-  updatedAt?: Maybe<Scalars['String']>;
-  createdBy?: Maybe<Scalars['String']>;
-  updatedBy?: Maybe<Scalars['String']>;
-  deletedAt?: Maybe<Scalars['String']>;
-  newId?: Maybe<Scalars['JSON']>;
-  dv?: Maybe<Scalars['Int']>;
-  sender?: Maybe<Scalars['JSON']>;
-  history_date?: Maybe<Scalars['String']>;
-  history_action?: Maybe<NewsItemUserReadHistoryRecordHistoryActionType>;
-  history_id?: Maybe<Scalars['String']>;
-};
-
-export type NewsItemUserReadHistoryRecordCreateInput = {
-  newsItem?: Maybe<Scalars['String']>;
-  user?: Maybe<Scalars['String']>;
-  v?: Maybe<Scalars['Int']>;
-  createdAt?: Maybe<Scalars['String']>;
-  updatedAt?: Maybe<Scalars['String']>;
-  createdBy?: Maybe<Scalars['String']>;
-  updatedBy?: Maybe<Scalars['String']>;
-  deletedAt?: Maybe<Scalars['String']>;
-  newId?: Maybe<Scalars['JSON']>;
-  dv?: Maybe<Scalars['Int']>;
-  sender?: Maybe<Scalars['JSON']>;
-  history_date?: Maybe<Scalars['String']>;
-  history_action?: Maybe<NewsItemUserReadHistoryRecordHistoryActionType>;
-  history_id?: Maybe<Scalars['String']>;
-};
-
-export enum NewsItemUserReadHistoryRecordHistoryActionType {
-  C = 'c',
-  U = 'u',
-  D = 'd'
-}
-
-export type NewsItemUserReadHistoryRecordUpdateInput = {
-  newsItem?: Maybe<Scalars['String']>;
-  user?: Maybe<Scalars['String']>;
-  v?: Maybe<Scalars['Int']>;
-  createdAt?: Maybe<Scalars['String']>;
-  updatedAt?: Maybe<Scalars['String']>;
-  createdBy?: Maybe<Scalars['String']>;
-  updatedBy?: Maybe<Scalars['String']>;
-  deletedAt?: Maybe<Scalars['String']>;
-  newId?: Maybe<Scalars['JSON']>;
-  dv?: Maybe<Scalars['Int']>;
-  sender?: Maybe<Scalars['JSON']>;
-  history_date?: Maybe<Scalars['String']>;
-  history_action?: Maybe<NewsItemUserReadHistoryRecordHistoryActionType>;
-  history_id?: Maybe<Scalars['String']>;
-};
-
-export type NewsItemUserReadHistoryRecordWhereInput = {
-  AND?: Maybe<Array<Maybe<NewsItemUserReadHistoryRecordWhereInput>>>;
-  OR?: Maybe<Array<Maybe<NewsItemUserReadHistoryRecordWhereInput>>>;
-  newsItem?: Maybe<Scalars['String']>;
-  newsItem_not?: Maybe<Scalars['String']>;
-  newsItem_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  newsItem_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  user?: Maybe<Scalars['String']>;
-  user_not?: Maybe<Scalars['String']>;
-  user_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  user_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  id?: Maybe<Scalars['ID']>;
-  id_not?: Maybe<Scalars['ID']>;
-  id_in?: Maybe<Array<Maybe<Scalars['ID']>>>;
-  id_not_in?: Maybe<Array<Maybe<Scalars['ID']>>>;
-  v?: Maybe<Scalars['Int']>;
-  v_not?: Maybe<Scalars['Int']>;
-  v_lt?: Maybe<Scalars['Int']>;
-  v_lte?: Maybe<Scalars['Int']>;
-  v_gt?: Maybe<Scalars['Int']>;
-  v_gte?: Maybe<Scalars['Int']>;
-  v_in?: Maybe<Array<Maybe<Scalars['Int']>>>;
-  v_not_in?: Maybe<Array<Maybe<Scalars['Int']>>>;
-  createdAt?: Maybe<Scalars['String']>;
-  createdAt_not?: Maybe<Scalars['String']>;
-  createdAt_lt?: Maybe<Scalars['String']>;
-  createdAt_lte?: Maybe<Scalars['String']>;
-  createdAt_gt?: Maybe<Scalars['String']>;
-  createdAt_gte?: Maybe<Scalars['String']>;
-  createdAt_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  createdAt_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  updatedAt?: Maybe<Scalars['String']>;
-  updatedAt_not?: Maybe<Scalars['String']>;
-  updatedAt_lt?: Maybe<Scalars['String']>;
-  updatedAt_lte?: Maybe<Scalars['String']>;
-  updatedAt_gt?: Maybe<Scalars['String']>;
-  updatedAt_gte?: Maybe<Scalars['String']>;
-  updatedAt_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  updatedAt_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  createdBy?: Maybe<Scalars['String']>;
-  createdBy_not?: Maybe<Scalars['String']>;
-  createdBy_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  createdBy_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  updatedBy?: Maybe<Scalars['String']>;
-  updatedBy_not?: Maybe<Scalars['String']>;
-  updatedBy_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  updatedBy_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  deletedAt?: Maybe<Scalars['String']>;
-  deletedAt_not?: Maybe<Scalars['String']>;
-  deletedAt_lt?: Maybe<Scalars['String']>;
-  deletedAt_lte?: Maybe<Scalars['String']>;
-  deletedAt_gt?: Maybe<Scalars['String']>;
-  deletedAt_gte?: Maybe<Scalars['String']>;
-  deletedAt_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  deletedAt_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  newId?: Maybe<Scalars['JSON']>;
-  newId_not?: Maybe<Scalars['JSON']>;
-  newId_in?: Maybe<Array<Maybe<Scalars['JSON']>>>;
-  newId_not_in?: Maybe<Array<Maybe<Scalars['JSON']>>>;
-  dv?: Maybe<Scalars['Int']>;
-  dv_not?: Maybe<Scalars['Int']>;
-  dv_lt?: Maybe<Scalars['Int']>;
-  dv_lte?: Maybe<Scalars['Int']>;
-  dv_gt?: Maybe<Scalars['Int']>;
-  dv_gte?: Maybe<Scalars['Int']>;
-  dv_in?: Maybe<Array<Maybe<Scalars['Int']>>>;
-  dv_not_in?: Maybe<Array<Maybe<Scalars['Int']>>>;
-  sender?: Maybe<Scalars['JSON']>;
-  sender_not?: Maybe<Scalars['JSON']>;
-  sender_in?: Maybe<Array<Maybe<Scalars['JSON']>>>;
-  sender_not_in?: Maybe<Array<Maybe<Scalars['JSON']>>>;
-  history_date?: Maybe<Scalars['String']>;
-  history_date_not?: Maybe<Scalars['String']>;
-  history_date_lt?: Maybe<Scalars['String']>;
-  history_date_lte?: Maybe<Scalars['String']>;
-  history_date_gt?: Maybe<Scalars['String']>;
-  history_date_gte?: Maybe<Scalars['String']>;
-  history_date_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  history_date_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  history_action?: Maybe<NewsItemUserReadHistoryRecordHistoryActionType>;
-  history_action_not?: Maybe<NewsItemUserReadHistoryRecordHistoryActionType>;
-  history_action_in?: Maybe<Array<Maybe<NewsItemUserReadHistoryRecordHistoryActionType>>>;
-  history_action_not_in?: Maybe<Array<Maybe<NewsItemUserReadHistoryRecordHistoryActionType>>>;
-  history_id?: Maybe<Scalars['String']>;
-  history_id_not?: Maybe<Scalars['String']>;
-  history_id_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  history_id_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-};
-
-export type NewsItemUserReadHistoryRecordWhereUniqueInput = {
-  id: Scalars['ID'];
-};
-
-export type NewsItemUserReadHistoryRecordsCreateInput = {
-  data?: Maybe<NewsItemUserReadHistoryRecordCreateInput>;
-};
-
-export type NewsItemUserReadHistoryRecordsUpdateInput = {
-  id: Scalars['ID'];
-  data?: Maybe<NewsItemUserReadHistoryRecordUpdateInput>;
-};
-
-export type NewsItemUserReadUpdateInput = {
-  newsItem?: Maybe<NewsItemRelateToOneInput>;
-  user?: Maybe<UserRelateToOneInput>;
-  v?: Maybe<Scalars['Int']>;
-  createdAt?: Maybe<Scalars['String']>;
-  updatedAt?: Maybe<Scalars['String']>;
-  createdBy?: Maybe<UserRelateToOneInput>;
-  updatedBy?: Maybe<UserRelateToOneInput>;
-  deletedAt?: Maybe<Scalars['String']>;
-  newId?: Maybe<Scalars['String']>;
-  dv?: Maybe<Scalars['Int']>;
-  sender?: Maybe<SenderFieldInput>;
-};
-
-export type NewsItemUserReadWhereInput = {
-  AND?: Maybe<Array<Maybe<NewsItemUserReadWhereInput>>>;
-  OR?: Maybe<Array<Maybe<NewsItemUserReadWhereInput>>>;
-  newsItem?: Maybe<NewsItemWhereInput>;
-  newsItem_is_null?: Maybe<Scalars['Boolean']>;
-  user?: Maybe<UserWhereInput>;
-  user_is_null?: Maybe<Scalars['Boolean']>;
-  id?: Maybe<Scalars['ID']>;
-  id_not?: Maybe<Scalars['ID']>;
-  id_in?: Maybe<Array<Maybe<Scalars['ID']>>>;
-  id_not_in?: Maybe<Array<Maybe<Scalars['ID']>>>;
-  v?: Maybe<Scalars['Int']>;
-  v_not?: Maybe<Scalars['Int']>;
-  v_lt?: Maybe<Scalars['Int']>;
-  v_lte?: Maybe<Scalars['Int']>;
-  v_gt?: Maybe<Scalars['Int']>;
-  v_gte?: Maybe<Scalars['Int']>;
-  v_in?: Maybe<Array<Maybe<Scalars['Int']>>>;
-  v_not_in?: Maybe<Array<Maybe<Scalars['Int']>>>;
-  createdAt?: Maybe<Scalars['String']>;
-  createdAt_not?: Maybe<Scalars['String']>;
-  createdAt_lt?: Maybe<Scalars['String']>;
-  createdAt_lte?: Maybe<Scalars['String']>;
-  createdAt_gt?: Maybe<Scalars['String']>;
-  createdAt_gte?: Maybe<Scalars['String']>;
-  createdAt_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  createdAt_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  updatedAt?: Maybe<Scalars['String']>;
-  updatedAt_not?: Maybe<Scalars['String']>;
-  updatedAt_lt?: Maybe<Scalars['String']>;
-  updatedAt_lte?: Maybe<Scalars['String']>;
-  updatedAt_gt?: Maybe<Scalars['String']>;
-  updatedAt_gte?: Maybe<Scalars['String']>;
-  updatedAt_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  updatedAt_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  createdBy?: Maybe<UserWhereInput>;
-  createdBy_is_null?: Maybe<Scalars['Boolean']>;
-  updatedBy?: Maybe<UserWhereInput>;
-  updatedBy_is_null?: Maybe<Scalars['Boolean']>;
-  deletedAt?: Maybe<Scalars['String']>;
-  deletedAt_not?: Maybe<Scalars['String']>;
-  deletedAt_lt?: Maybe<Scalars['String']>;
-  deletedAt_lte?: Maybe<Scalars['String']>;
-  deletedAt_gt?: Maybe<Scalars['String']>;
-  deletedAt_gte?: Maybe<Scalars['String']>;
-  deletedAt_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  deletedAt_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  newId?: Maybe<Scalars['String']>;
-  newId_not?: Maybe<Scalars['String']>;
-  newId_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  newId_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  dv?: Maybe<Scalars['Int']>;
-  dv_not?: Maybe<Scalars['Int']>;
-  dv_lt?: Maybe<Scalars['Int']>;
-  dv_lte?: Maybe<Scalars['Int']>;
-  dv_gt?: Maybe<Scalars['Int']>;
-  dv_gte?: Maybe<Scalars['Int']>;
-  dv_in?: Maybe<Array<Maybe<Scalars['Int']>>>;
-  dv_not_in?: Maybe<Array<Maybe<Scalars['Int']>>>;
-  sender?: Maybe<SenderFieldInput>;
-  sender_not?: Maybe<SenderFieldInput>;
-  sender_in?: Maybe<Array<Maybe<SenderFieldInput>>>;
-  sender_not_in?: Maybe<Array<Maybe<SenderFieldInput>>>;
-};
-
-export type NewsItemUserReadWhereUniqueInput = {
-  id: Scalars['ID'];
-};
-
-export type NewsItemUserReadsCreateInput = {
-  data?: Maybe<NewsItemUserReadCreateInput>;
-};
-
-export type NewsItemUserReadsUpdateInput = {
-  id: Scalars['ID'];
-  data?: Maybe<NewsItemUserReadUpdateInput>;
 };
 
 export type NewsItemWhereInput = {
@@ -47245,22 +46852,6 @@ export type Query = {
   _allNewsItemTemplatesMeta?: Maybe<_QueryMeta>;
   /**  Retrieve the meta-data for the NewsItemTemplate list.  */
   _NewsItemTemplatesMeta?: Maybe<_ListMeta>;
-  /**  Search for all NewsItemUserReadHistoryRecord items which match the where clause.  */
-  allNewsItemUserReadHistoryRecords?: Maybe<Array<Maybe<NewsItemUserReadHistoryRecord>>>;
-  /**  Search for the NewsItemUserReadHistoryRecord item with the matching ID.  */
-  NewsItemUserReadHistoryRecord?: Maybe<NewsItemUserReadHistoryRecord>;
-  /**  Perform a meta-query on all NewsItemUserReadHistoryRecord items which match the where clause.  */
-  _allNewsItemUserReadHistoryRecordsMeta?: Maybe<_QueryMeta>;
-  /**  Retrieve the meta-data for the NewsItemUserReadHistoryRecord list.  */
-  _NewsItemUserReadHistoryRecordsMeta?: Maybe<_ListMeta>;
-  /**  Search for all NewsItemUserRead items which match the where clause.  */
-  allNewsItemUserReads?: Maybe<Array<Maybe<NewsItemUserRead>>>;
-  /**  Search for the NewsItemUserRead item with the matching ID.  */
-  NewsItemUserRead?: Maybe<NewsItemUserRead>;
-  /**  Perform a meta-query on all NewsItemUserRead items which match the where clause.  */
-  _allNewsItemUserReadsMeta?: Maybe<_QueryMeta>;
-  /**  Retrieve the meta-data for the NewsItemUserRead list.  */
-  _NewsItemUserReadsMeta?: Maybe<_ListMeta>;
   /**  Search for all WebhookHistoryRecord items which match the where clause.  */
   allWebhookHistoryRecords?: Maybe<Array<Maybe<WebhookHistoryRecord>>>;
   /**  Search for the WebhookHistoryRecord item with the matching ID.  */
@@ -52365,56 +51956,6 @@ export type Query_AllNewsItemTemplatesMetaArgs = {
   where?: Maybe<NewsItemTemplateWhereInput>;
   search?: Maybe<Scalars['String']>;
   sortBy?: Maybe<Array<SortNewsItemTemplatesBy>>;
-  orderBy?: Maybe<Scalars['String']>;
-  first?: Maybe<Scalars['Int']>;
-  skip?: Maybe<Scalars['Int']>;
-};
-
-
-export type QueryAllNewsItemUserReadHistoryRecordsArgs = {
-  where?: Maybe<NewsItemUserReadHistoryRecordWhereInput>;
-  search?: Maybe<Scalars['String']>;
-  sortBy?: Maybe<Array<SortNewsItemUserReadHistoryRecordsBy>>;
-  orderBy?: Maybe<Scalars['String']>;
-  first?: Maybe<Scalars['Int']>;
-  skip?: Maybe<Scalars['Int']>;
-};
-
-
-export type QueryNewsItemUserReadHistoryRecordArgs = {
-  where: NewsItemUserReadHistoryRecordWhereUniqueInput;
-};
-
-
-export type Query_AllNewsItemUserReadHistoryRecordsMetaArgs = {
-  where?: Maybe<NewsItemUserReadHistoryRecordWhereInput>;
-  search?: Maybe<Scalars['String']>;
-  sortBy?: Maybe<Array<SortNewsItemUserReadHistoryRecordsBy>>;
-  orderBy?: Maybe<Scalars['String']>;
-  first?: Maybe<Scalars['Int']>;
-  skip?: Maybe<Scalars['Int']>;
-};
-
-
-export type QueryAllNewsItemUserReadsArgs = {
-  where?: Maybe<NewsItemUserReadWhereInput>;
-  search?: Maybe<Scalars['String']>;
-  sortBy?: Maybe<Array<SortNewsItemUserReadsBy>>;
-  orderBy?: Maybe<Scalars['String']>;
-  first?: Maybe<Scalars['Int']>;
-  skip?: Maybe<Scalars['Int']>;
-};
-
-
-export type QueryNewsItemUserReadArgs = {
-  where: NewsItemUserReadWhereUniqueInput;
-};
-
-
-export type Query_AllNewsItemUserReadsMetaArgs = {
-  where?: Maybe<NewsItemUserReadWhereInput>;
-  search?: Maybe<Scalars['String']>;
-  sortBy?: Maybe<Array<SortNewsItemUserReadsBy>>;
   orderBy?: Maybe<Scalars['String']>;
   first?: Maybe<Scalars['Int']>;
   skip?: Maybe<Scalars['Int']>;
@@ -59432,48 +58973,6 @@ export enum SortNewsItemTemplatesBy {
   TitleDesc = 'title_DESC',
   BodyAsc = 'body_ASC',
   BodyDesc = 'body_DESC',
-  IdAsc = 'id_ASC',
-  IdDesc = 'id_DESC',
-  VAsc = 'v_ASC',
-  VDesc = 'v_DESC',
-  CreatedAtAsc = 'createdAt_ASC',
-  CreatedAtDesc = 'createdAt_DESC',
-  UpdatedAtAsc = 'updatedAt_ASC',
-  UpdatedAtDesc = 'updatedAt_DESC',
-  CreatedByAsc = 'createdBy_ASC',
-  CreatedByDesc = 'createdBy_DESC',
-  UpdatedByAsc = 'updatedBy_ASC',
-  UpdatedByDesc = 'updatedBy_DESC',
-  DeletedAtAsc = 'deletedAt_ASC',
-  DeletedAtDesc = 'deletedAt_DESC',
-  DvAsc = 'dv_ASC',
-  DvDesc = 'dv_DESC'
-}
-
-export enum SortNewsItemUserReadHistoryRecordsBy {
-  IdAsc = 'id_ASC',
-  IdDesc = 'id_DESC',
-  VAsc = 'v_ASC',
-  VDesc = 'v_DESC',
-  CreatedAtAsc = 'createdAt_ASC',
-  CreatedAtDesc = 'createdAt_DESC',
-  UpdatedAtAsc = 'updatedAt_ASC',
-  UpdatedAtDesc = 'updatedAt_DESC',
-  DeletedAtAsc = 'deletedAt_ASC',
-  DeletedAtDesc = 'deletedAt_DESC',
-  DvAsc = 'dv_ASC',
-  DvDesc = 'dv_DESC',
-  HistoryDateAsc = 'history_date_ASC',
-  HistoryDateDesc = 'history_date_DESC',
-  HistoryActionAsc = 'history_action_ASC',
-  HistoryActionDesc = 'history_action_DESC'
-}
-
-export enum SortNewsItemUserReadsBy {
-  NewsItemAsc = 'newsItem_ASC',
-  NewsItemDesc = 'newsItem_DESC',
-  UserAsc = 'user_ASC',
-  UserDesc = 'user_DESC',
   IdAsc = 'id_ASC',
   IdDesc = 'id_DESC',
   VAsc = 'v_ASC',
